pyproject.toml
[project]
name = "track-job-hackathon-03020301"
version = "0.1.0"
description = ""
authors = [
    {name = "Conan-Wen",email = "w245646037@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.65.1,<2.0.0)",
    "ics (>=0.7.2,<0.8.0)",
    "streamlit (>=1.42.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "track-job-hackathon-0302", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
